requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite
discord.py
python-dotenv
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1
asyncio==3.4.3
youtube_dl
ffmpeg